Added Hibernate support to TagDAO

+++ restbasic-app/src/main/java/com/epam/esm/bahlei/restbasic/model/Pageable.java
@@ -1,27 +1,27 @@
 package com.epam.esm.bahlei.restbasic.model;
 
 public class Pageable {
-  private long offset;
-  private long limit;
-  private long page;
+  private int offset;
+  private int limit;
+  private int page;
 
   public Pageable() {}
 
-  public Pageable(long page, long limit) {
+  public Pageable(int page, int limit) {
     this.limit = limit > 0 ? limit : 0;
     this.page = page > 0 ? page : 1;
     this.offset = this.limit * (this.page - 1);
   }
 
-  public long getOffset() {
+  public int getOffset() {
     return offset;
   }
 
-  public long getLimit() {
+  public int getLimit() {
     return limit;
   }
 
-  public long getPage() {
+  public int getPage() {
     return page;
   }
 }
